Introduce factory pattern for Mediator
