+++ mediator/src/main/java/cas/ypsiliform/mediator/agentregistration/AgentRegistrationBean.java
@@ -22,6 +22,8 @@ import javax.enterprise.concurrent.ManagedExecutorService;
 import javax.enterprise.event.Observes;
 
 import cas.ypsiliform.mediator.Mediator;
+import cas.ypsiliform.mediator.MediatorFactory;
+import cas.ypsiliform.mediator.VoteBasedMediator;
 import cas.ypsiliform.mediator.negotiation.AgentProxy;
 import cas.ypsiliform.mediator.websocket.NewMessageEvent;
 import cas.ypsiliform.mediator.websocket.SessionRepository;
@@ -85,10 +87,9 @@ public class AgentRegistrationBean
             {
                 LOGGER.log(Level.FINE, "5 agent are registered start mediator");
                 Integer lowestKey = Collections.min(agentMap.keySet());
-                Mediator mediator = new Mediator(agentMap,
-                                                 agentMap.get(lowestKey)
-                                                     .getAgentData()
-                                                     .getInitialDemand());
+                
+                Mediator mediator = MediatorFactory.getMediator(
+                		MediatorFactory.Strategy.VOTING, agentMap, agentMap.get(lowestKey).getAgentData().getInitialDemand());
                 service.execute(mediator);
             }
         }
